+++ lib/mix/tasks/mishka.ui.add.ex
@@ -167,6 +167,7 @@ defmodule Mix.Tasks.Mishka.Ui.Add do
     %Igniter.Mix.Task.Args{positional: %{repo: repo}} = igniter.args
 
     options = igniter.args.options
+    tty? = IO.ANSI.enabled?()
 
     if !options[:test] do
       msg =
@@ -181,7 +182,7 @@ defmodule Mix.Tasks.Mishka.Ui.Add do
       IO.puts(IO.ANSI.blue() <> String.trim_trailing(msg) <> IO.ANSI.reset())
     end
 
-    if !options[:test],
+    if !options[:test] and tty?,
       do: Owl.Spinner.start(id: :my_spinner, labels: [processing: "Please wait..."])
 
     {url, repo_action, igniter} =
@@ -248,7 +249,7 @@ defmodule Mix.Tasks.Mishka.Ui.Add do
         end
       end
 
-    if !options[:test] do
+    if !options[:test] and tty? do
       if Map.get(final_igniter, :issues, []) == [],
         do: Owl.Spinner.stop(id: :my_spinner, resolution: :ok, label: "Done"),
         else: Owl.Spinner.stop(id: :my_spinner, resolution: :error, label: "Error")
@@ -406,17 +407,18 @@ defmodule Mix.Tasks.Mishka.Ui.Add do
 
   defp convert_request_body(body, url, igniter) when url in @domain_types do
     options = igniter.args.options
-    if !options[:test], do: Owl.Spinner.stop(id: :my_spinner, resolution: :ok)
+    tty? = IO.ANSI.enabled?()
+    if !options[:test] and tty?, do: Owl.Spinner.stop(id: :my_spinner, resolution: :ok)
 
     case Igniter.Util.IO.yes?(external_call_warning()) do
       false ->
-        if !options[:test],
+        if !options[:test] and tty?,
           do: Owl.Spinner.start(id: :my_spinner, labels: [processing: "Please wait..."])
 
         {:error, "The operation was stopped at your request."}
 
       true ->
-        if !options[:test],
+        if !options[:test] and tty?,
           do: Owl.Spinner.start(id: :my_spinner, labels: [processing: "Please wait..."])
 
         if url == :github,